requirements.txt
flask
flask-cors
python-dotenv
requests
chromadb
sentence-transformers
numpy
pandas
beautifulsoup4
black
isort